Refactoring Hessian evaluation
