+++ examples/examples/ru1121_structures/c.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:24.000288Z'
+  generated: '2026-03-16T14:10:25.238045Z'
 structure:
   formula: Ru48C2
   n_atoms: 50
@@ -65,6 +65,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 6.5.1
+  executed_at: '2025-11-02T10:24:49Z'
   type: frequency
   incar:
     LREAL: T
@@ -102,40 +104,40 @@ vibrations:
     - 50Y
     - 50Z
     matrix:
-    - - -11.028369
-      - -0.476227
-      - -1.571812
-      - -0.009702
-      - 0.016878
-      - -0.015082
-    - - -0.471404
-      - -11.996908
-      - 0.196398
-      - 0.008667
-      - -0.029319
-      - 0.007914
-    - - -1.588191
-      - 0.277289
-      - -6.069597
-      - -0.005006
-      - 0.018083
-      - 0.021776
-    - - -0.004455
-      - 0.018706
-      - -0.081196
-      - -11.30584
-      - -0.341375
-      - -1.388921
-    - - 0.008089
-      - -0.019896
-      - 0.001058
-      - -0.336609
-      - -12.542438
-      - 0.561785
-    - - -0.007441
-      - 0.025501
-      - 0.02044
-      - -1.324243
-      - -0.032636
-      - -5.942638
+    - - 11.028369
+      - 0.476227
+      - 1.571812
+      - 0.009702
+      - -0.016878
+      - 0.015082
+    - - 0.471404
+      - 11.996908
+      - -0.196398
+      - -0.008667
+      - 0.029319
+      - -0.007914
+    - - 1.588191
+      - -0.277289
+      - 6.069597
+      - 0.005006
+      - -0.018083
+      - -0.021776
+    - - 0.004455
+      - -0.018706
+      - 0.081196
+      - 11.30584
+      - 0.341375
+      - 1.388921
+    - - -0.008089
+      - 0.019896
+      - -0.001058
+      - 0.336609
+      - 12.542438
+      - -0.561785
+    - - 0.007441
+      - -0.025501
+      - -0.02044
+      - 1.324243
+      - 0.032636
+      - 5.942638
   paired_modes_averaged: true
