+++ examples/examples/ru1121_structures/c_hydr.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:04:47.341822Z'
+  generated: '2026-03-16T14:10:36.110402Z'
 structure:
   formula: Ru48C2H2
   n_atoms: 52
@@ -67,6 +67,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 4.6.31
+  executed_at: '2012-12-01T23:51:38Z'
   type: frequency
   incar:
     PREC: medium
@@ -114,150 +116,150 @@ vibrations:
     - 52Y
     - 52Z
     matrix:
-    - - -0.303555
-      - 0.146217
-      - -0.17347
-      - -0.000603
-      - 0.002556
-      - -0.001884
-      - -1.251531
-      - -0.515041
-      - -0.182754
-      - 0.003997
-      - 0.004958
-      - 0.001161
-    - - 0.146217
-      - -0.659448
-      - -0.086152
-      - 0.000495
-      - -0.002879
-      - 0.000879
-      - -0.45793
-      - -0.086379
-      - -0.021873
-      - 0.002162
-      - 0.000797
-      - -0.002227
-    - - -0.17347
-      - -0.086152
-      - -0.610553
-      - 0.000347
-      - 0.001267
-      - -0.000142
-      - 0.691764
-      - 0.293793
-      - 0.124043
-      - -0.000118
-      - 0.004665
-      - 0.004744
-    - - -0.000603
-      - 0.000495
-      - 0.000347
-      - -1.04929
-      - -0.029243
-      - -0.232422
-      - 0.002928
-      - -0.000354
-      - 0.001233
-      - 0.049979
-      - 0.021978
-      - -0.012843
-    - - 0.002556
-      - -0.002879
-      - 0.001267
-      - -0.029243
-      - -1.199286
-      - -0.00555
-      - -0.00065
-      - 0.005866
-      - -0.002624
-      - 0.008317
-      - 0.00214
-      - 0.009466
-    - - -0.001884
-      - 0.000879
-      - -0.000142
-      - -0.232422
-      - -0.00555
-      - -0.383341
-      - 0.004899
-      - 0.003818
-      - 0.002277
-      - -0.01849
-      - -0.004611
-      - -0.00306
-    - - -1.251531
-      - -0.45793
-      - 0.691764
-      - 0.002928
-      - -0.00065
-      - 0.004899
-      - 2.545268
-      - 0.075101
-      - -0.863998
-      - -0.008063
-      - -0.010546
-      - -0.003323
-    - - -0.515041
-      - -0.086379
-      - 0.293793
-      - -0.000354
-      - 0.005866
-      - 0.003818
-      - 0.075101
-      - -1.420944
-      - -1.120221
-      - -0.000969
-      - -0.009406
-      - 0.004048
-    - - -0.182754
-      - -0.021873
-      - 0.124043
-      - 0.001233
-      - -0.002624
-      - 0.002277
-      - -0.863998
-      - -1.120221
-      - -2.837667
-      - 0.003604
-      - -0.005831
-      - 0.00111
-    - - 0.003997
-      - 0.002162
-      - -0.000118
-      - 0.049979
-      - 0.008317
-      - -0.01849
-      - -0.008063
-      - -0.000969
-      - 0.003604
-      - -0.553946
-      - 0.82009
-      - -0.249756
-    - - 0.004958
-      - 0.000797
-      - 0.004665
-      - 0.021978
-      - 0.00214
-      - -0.004611
-      - -0.010546
-      - -0.009406
-      - -0.005831
-      - 0.82009
-      - -3.935401
-      - -0.752936
-    - - 0.001161
-      - -0.002227
-      - 0.004744
-      - -0.012843
-      - 0.009466
-      - -0.00306
-      - -0.003323
-      - 0.004048
-      - 0.00111
-      - -0.249756
-      - -0.752936
-      - -5.801363
+    - - 0.303555
+      - -0.146217
+      - 0.17347
+      - 0.000603
+      - -0.002556
+      - 0.001884
+      - 1.251531
+      - 0.515041
+      - 0.182754
+      - -0.003997
+      - -0.004958
+      - -0.001161
+    - - -0.146217
+      - 0.659448
+      - 0.086152
+      - -0.000495
+      - 0.002879
+      - -0.000879
+      - 0.45793
+      - 0.086379
+      - 0.021873
+      - -0.002162
+      - -0.000797
+      - 0.002227
+    - - 0.17347
+      - 0.086152
+      - 0.610553
+      - -0.000347
+      - -0.001267
+      - 0.000142
+      - -0.691764
+      - -0.293793
+      - -0.124043
+      - 0.000118
+      - -0.004665
+      - -0.004744
+    - - 0.000603
+      - -0.000495
+      - -0.000347
+      - 1.04929
+      - 0.029243
+      - 0.232422
+      - -0.002928
+      - 0.000354
+      - -0.001233
+      - -0.049979
+      - -0.021978
+      - 0.012843
+    - - -0.002556
+      - 0.002879
+      - -0.001267
+      - 0.029243
+      - 1.199286
+      - 0.00555
+      - 0.00065
+      - -0.005866
+      - 0.002624
+      - -0.008317
+      - -0.00214
+      - -0.009466
+    - - 0.001884
+      - -0.000879
+      - 0.000142
+      - 0.232422
+      - 0.00555
+      - 0.383341
+      - -0.004899
+      - -0.003818
+      - -0.002277
+      - 0.01849
+      - 0.004611
+      - 0.00306
+    - - 1.251531
+      - 0.45793
+      - -0.691764
+      - -0.002928
+      - 0.00065
+      - -0.004899
+      - -2.545268
+      - -0.075101
+      - 0.863998
+      - 0.008063
+      - 0.010546
+      - 0.003323
+    - - 0.515041
+      - 0.086379
+      - -0.293793
+      - 0.000354
+      - -0.005866
+      - -0.003818
+      - -0.075101
+      - 1.420944
+      - 1.120221
+      - 0.000969
+      - 0.009406
+      - -0.004048
+    - - 0.182754
+      - 0.021873
+      - -0.124043
+      - -0.001233
+      - 0.002624
+      - -0.002277
+      - 0.863998
+      - 1.120221
+      - 2.837667
+      - -0.003604
+      - 0.005831
+      - -0.00111
+    - - -0.003997
+      - -0.002162
+      - 0.000118
+      - -0.049979
+      - -0.008317
+      - 0.01849
+      - 0.008063
+      - 0.000969
+      - -0.003604
+      - 0.553946
+      - -0.82009
+      - 0.249756
+    - - -0.004958
+      - -0.000797
+      - -0.004665
+      - -0.021978
+      - -0.00214
+      - 0.004611
+      - 0.010546
+      - 0.009406
+      - 0.005831
+      - -0.82009
+      - 3.935401
+      - 0.752936
+    - - -0.001161
+      - 0.002227
+      - -0.004744
+      - 0.012843
+      - -0.009466
+      - 0.00306
+      - 0.003323
+      - -0.004048
+      - -0.00111
+      - 0.249756
+      - 0.752936
+      - 5.801363
   paired_modes_averaged: true
   pairing_note: Odd number of real modes with one imaginary mode (TS case). Dropped
     unpaired real mode (278.325652 cm-1) before pair averaging.
