+++ examples/examples/ru1121_structures/ch.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:22.503823Z'
+  generated: '2026-03-16T14:10:25.971933Z'
 structure:
   formula: Ru48C2H2
   n_atoms: 52
@@ -67,6 +67,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 4.6.31
+  executed_at: '2012-11-02T09:49:06Z'
   type: frequency
   incar:
     PREC: medium
@@ -113,148 +115,148 @@ vibrations:
     - 52Y
     - 52Z
     matrix:
-    - - -0.989017
-      - -0.139086
-      - 0.822749
-      - 0.001055
-      - 0.002479
-      - -6.5e-05
-      - 1.495105
-      - 0.326295
-      - -2.712961
-      - 0.002007
-      - -0.002328
-      - 0.002598
-    - - -0.139086
-      - -0.605948
-      - 0.12126
-      - 0.002404
-      - 0.000847
-      - -0.000411
-      - 0.317441
-      - 0.330244
-      - -0.54176
-      - -0.000116
-      - 0.001556
-      - 0.00353
-    - - 0.822749
-      - 0.12126
-      - -2.632011
-      - -0.003796
-      - -0.000479
-      - 0.004169
-      - -2.761539
-      - -0.548045
-      - 6.581537
-      - -0.003278
-      - -0.002872
-      - -0.001156
-    - - 0.001055
-      - 0.002404
-      - -0.003796
-      - -0.994272
-      - -0.138064
-      - 0.825162
-      - 0.001248
-      - -0.000789
-      - 0.007314
-      - 1.498537
-      - 0.325603
-      - -2.722331
-    - - 0.002479
-      - 0.000847
-      - -0.000479
-      - -0.138064
-      - -0.608038
-      - 0.120695
-      - 0.001383
-      - 0.003154
-      - 0.003491
-      - 0.319841
-      - 0.331119
-      - -0.540589
-    - - -6.5e-05
-      - -0.000411
-      - 0.004169
-      - 0.825162
-      - 0.120695
-      - -2.639441
-      - -0.019903
-      - -0.000259
-      - -0.000827
-      - -2.766701
-      - -0.547478
-      - 6.599722
-    - - 1.495105
-      - 0.317441
-      - -2.761539
-      - 0.001248
-      - 0.001383
-      - -0.019903
-      - -5.761988
-      - -0.837441
-      - 9.518617
-      - -0.007118
-      - 0.002095
-      - 0.017923
-    - - 0.326295
-      - 0.330244
-      - -0.548045
-      - -0.000789
-      - 0.003154
-      - -0.000259
-      - -0.837441
-      - -1.919625
-      - 1.950967
-      - 0.002344
-      - -0.000304
-      - 0.001916
-    - - -2.712961
-      - -0.54176
-      - 6.581537
-      - 0.007314
-      - 0.003491
-      - -0.000827
-      - 9.518617
-      - 1.950967
-      - -24.172054
-      - 0.003879
-      - 0.005471
-      - 0.003477
-    - - 0.002007
-      - -0.000116
-      - -0.003278
-      - 1.498537
-      - 0.319841
-      - -2.766701
-      - -0.007118
-      - 0.002344
-      - 0.003879
-      - -5.789883
-      - -0.81443
-      - 9.52815
-    - - -0.002328
-      - 0.001556
-      - -0.002872
-      - 0.325603
-      - 0.331119
-      - -0.547478
-      - 0.002095
-      - -0.000304
-      - 0.005471
-      - -0.81443
-      - -1.939706
-      - 1.93282
-    - - 0.002598
-      - 0.00353
-      - -0.001156
-      - -2.722331
-      - -0.540589
-      - 6.599722
-      - 0.017923
-      - 0.001916
-      - 0.003477
-      - 9.52815
-      - 1.93282
-      - -24.241201
+    - - 0.989017
+      - 0.139086
+      - -0.822749
+      - -0.001055
+      - -0.002479
+      - 6.5e-05
+      - -1.495105
+      - -0.326295
+      - 2.712961
+      - -0.002007
+      - 0.002328
+      - -0.002598
+    - - 0.139086
+      - 0.605948
+      - -0.12126
+      - -0.002404
+      - -0.000847
+      - 0.000411
+      - -0.317441
+      - -0.330244
+      - 0.54176
+      - 0.000116
+      - -0.001556
+      - -0.00353
+    - - -0.822749
+      - -0.12126
+      - 2.632011
+      - 0.003796
+      - 0.000479
+      - -0.004169
+      - 2.761539
+      - 0.548045
+      - -6.581537
+      - 0.003278
+      - 0.002872
+      - 0.001156
+    - - -0.001055
+      - -0.002404
+      - 0.003796
+      - 0.994272
+      - 0.138064
+      - -0.825162
+      - -0.001248
+      - 0.000789
+      - -0.007314
+      - -1.498537
+      - -0.325603
+      - 2.722331
+    - - -0.002479
+      - -0.000847
+      - 0.000479
+      - 0.138064
+      - 0.608038
+      - -0.120695
+      - -0.001383
+      - -0.003154
+      - -0.003491
+      - -0.319841
+      - -0.331119
+      - 0.540589
+    - - 6.5e-05
+      - 0.000411
+      - -0.004169
+      - -0.825162
+      - -0.120695
+      - 2.639441
+      - 0.019903
+      - 0.000259
+      - 0.000827
+      - 2.766701
+      - 0.547478
+      - -6.599722
+    - - -1.495105
+      - -0.317441
+      - 2.761539
+      - -0.001248
+      - -0.001383
+      - 0.019903
+      - 5.761988
+      - 0.837441
+      - -9.518617
+      - 0.007118
+      - -0.002095
+      - -0.017923
+    - - -0.326295
+      - -0.330244
+      - 0.548045
+      - 0.000789
+      - -0.003154
+      - 0.000259
+      - 0.837441
+      - 1.919625
+      - -1.950967
+      - -0.002344
+      - 0.000304
+      - -0.001916
+    - - 2.712961
+      - 0.54176
+      - -6.581537
+      - -0.007314
+      - -0.003491
+      - 0.000827
+      - -9.518617
+      - -1.950967
+      - 24.172054
+      - -0.003879
+      - -0.005471
+      - -0.003477
+    - - -0.002007
+      - 0.000116
+      - 0.003278
+      - -1.498537
+      - -0.319841
+      - 2.766701
+      - 0.007118
+      - -0.002344
+      - -0.003879
+      - 5.789883
+      - 0.81443
+      - -9.52815
+    - - 0.002328
+      - -0.001556
+      - 0.002872
+      - -0.325603
+      - -0.331119
+      - 0.547478
+      - -0.002095
+      - 0.000304
+      - -0.005471
+      - 0.81443
+      - 1.939706
+      - -1.93282
+    - - -0.002598
+      - -0.00353
+      - 0.001156
+      - 2.722331
+      - 0.540589
+      - -6.599722
+      - -0.017923
+      - -0.001916
+      - -0.003477
+      - -9.52815
+      - -1.93282
+      - 24.241201
   paired_modes_averaged: true
