+++ examples/examples/ru1121_structures/co__1.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:26.131999Z'
+  generated: '2026-03-16T14:10:28.370588Z'
 structure:
   formula: Ru48C2O2
   n_atoms: 52
@@ -67,6 +67,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 4.6.31
+  executed_at: '2012-11-02T09:49:07Z'
   type: frequency
   incar:
     PREC: medium
@@ -113,148 +115,148 @@ vibrations:
     - 52Y
     - 52Z
     matrix:
-    - - -1.580392
-      - -0.591409
-      - 1.12077
-      - 0.001675
-      - 0.00183
-      - -0.000878
-      - 0.838894
-      - 0.374156
-      - -1.085959
-      - -0.003056
-      - -0.001528
-      - -0.001183
-    - - -0.591409
-      - -0.627009
-      - 0.400151
-      - 0.001781
-      - 0.003104
-      - 0.001506
-      - 0.372826
-      - 0.26794
-      - -0.489451
-      - -0.000504
-      - -0.001319
-      - -0.000503
-    - - 1.12077
-      - 0.400151
-      - -1.648498
-      - -0.001663
-      - 0.001638
-      - 8.3e-05
-      - -0.745256
-      - -0.339975
-      - 1.262836
-      - 0.001501
-      - 0.000359
-      - -0.00028
-    - - 0.001675
-      - 0.001781
-      - -0.001663
-      - -1.568552
-      - -0.588057
-      - 1.108398
-      - -0.002467
-      - -0.001122
-      - -0.000133
-      - 0.831706
-      - 0.371131
-      - -1.075343
-    - - 0.00183
-      - 0.003104
-      - 0.001638
-      - -0.588057
-      - -0.627659
-      - 0.396204
-      - 0.000208
-      - -0.000919
-      - 0.000457
-      - 0.368929
-      - 0.268664
-      - -0.483078
-    - - -0.000878
-      - 0.001506
-      - 8.3e-05
-      - 1.108398
-      - 0.396204
-      - -1.6411
-      - -0.00054
-      - 0.000924
-      - 1.9e-05
-      - -0.735375
-      - -0.336589
-      - 1.254668
-    - - 0.838894
-      - 0.372826
-      - -0.745256
-      - -0.002467
-      - 0.000208
-      - -0.00054
-      - -0.794648
-      - -0.369277
-      - 0.846185
-      - -0.000734
-      - -0.001102
-      - 0.002502
-    - - 0.374156
-      - 0.26794
-      - -0.339975
-      - -0.001122
-      - -0.000919
-      - 0.000924
-      - -0.369277
-      - -0.325989
-      - 0.333149
-      - -0.000243
-      - 0.000662
-      - -0.000252
-    - - -1.085959
-      - -0.489451
-      - 1.262836
-      - -0.000133
-      - 0.000457
-      - 1.9e-05
-      - 0.846185
-      - 0.333149
-      - -1.556045
-      - 0.000412
-      - -0.000682
-      - 0.000452
-    - - -0.003056
-      - -0.000504
-      - 0.001501
-      - 0.831706
-      - 0.368929
-      - -0.735375
-      - -0.000734
-      - -0.000243
-      - 0.000412
-      - -0.784946
-      - -0.372879
-      - 0.838679
-    - - -0.001528
-      - -0.001319
-      - 0.000359
-      - 0.371131
-      - 0.268664
-      - -0.336589
-      - -0.001102
-      - 0.000662
-      - -0.000682
-      - -0.372879
-      - -0.330042
-      - 0.330674
-    - - -0.001183
-      - -0.000503
-      - -0.00028
-      - -1.075343
-      - -0.483078
-      - 1.254668
-      - 0.002502
-      - -0.000252
-      - 0.000452
-      - 0.838679
-      - 0.330674
-      - -1.551832
+    - - 1.580392
+      - 0.591409
+      - -1.12077
+      - -0.001675
+      - -0.00183
+      - 0.000878
+      - -0.838894
+      - -0.374156
+      - 1.085959
+      - 0.003056
+      - 0.001528
+      - 0.001183
+    - - 0.591409
+      - 0.627009
+      - -0.400151
+      - -0.001781
+      - -0.003104
+      - -0.001506
+      - -0.372826
+      - -0.26794
+      - 0.489451
+      - 0.000504
+      - 0.001319
+      - 0.000503
+    - - -1.12077
+      - -0.400151
+      - 1.648498
+      - 0.001663
+      - -0.001638
+      - -8.3e-05
+      - 0.745256
+      - 0.339975
+      - -1.262836
+      - -0.001501
+      - -0.000359
+      - 0.00028
+    - - -0.001675
+      - -0.001781
+      - 0.001663
+      - 1.568552
+      - 0.588057
+      - -1.108398
+      - 0.002467
+      - 0.001122
+      - 0.000133
+      - -0.831706
+      - -0.371131
+      - 1.075343
+    - - -0.00183
+      - -0.003104
+      - -0.001638
+      - 0.588057
+      - 0.627659
+      - -0.396204
+      - -0.000208
+      - 0.000919
+      - -0.000457
+      - -0.368929
+      - -0.268664
+      - 0.483078
+    - - 0.000878
+      - -0.001506
+      - -8.3e-05
+      - -1.108398
+      - -0.396204
+      - 1.6411
+      - 0.00054
+      - -0.000924
+      - -1.9e-05
+      - 0.735375
+      - 0.336589
+      - -1.254668
+    - - -0.838894
+      - -0.372826
+      - 0.745256
+      - 0.002467
+      - -0.000208
+      - 0.00054
+      - 0.794648
+      - 0.369277
+      - -0.846185
+      - 0.000734
+      - 0.001102
+      - -0.002502
+    - - -0.374156
+      - -0.26794
+      - 0.339975
+      - 0.001122
+      - 0.000919
+      - -0.000924
+      - 0.369277
+      - 0.325989
+      - -0.333149
+      - 0.000243
+      - -0.000662
+      - 0.000252
+    - - 1.085959
+      - 0.489451
+      - -1.262836
+      - 0.000133
+      - -0.000457
+      - -1.9e-05
+      - -0.846185
+      - -0.333149
+      - 1.556045
+      - -0.000412
+      - 0.000682
+      - -0.000452
+    - - 0.003056
+      - 0.000504
+      - -0.001501
+      - -0.831706
+      - -0.368929
+      - 0.735375
+      - 0.000734
+      - 0.000243
+      - -0.000412
+      - 0.784946
+      - 0.372879
+      - -0.838679
+    - - 0.001528
+      - 0.001319
+      - -0.000359
+      - -0.371131
+      - -0.268664
+      - 0.336589
+      - 0.001102
+      - -0.000662
+      - 0.000682
+      - 0.372879
+      - 0.330042
+      - -0.330674
+    - - 0.001183
+      - 0.000503
+      - 0.00028
+      - 1.075343
+      - 0.483078
+      - -1.254668
+      - -0.002502
+      - 0.000252
+      - -0.000452
+      - -0.838679
+      - -0.330674
+      - 1.551832
   paired_modes_averaged: true
