+++ examples/examples/ru1121_structures/co__2.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:32.100093Z'
+  generated: '2026-03-16T14:10:38.535765Z'
 structure:
   formula: CO
   n_atoms: 2
@@ -17,6 +17,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 6.5.0
+  executed_at: '2025-11-02T08:36:37Z'
   type: frequency
   incar:
     LREAL: F
@@ -58,39 +60,39 @@ vibrations:
     - 2Y
     - 2Z
     matrix:
-    - - -113.476632
+    - - 113.476632
+      - -0.0
+      - -0.0
+      - -113.476632
       - 0.0
       - 0.0
-      - 113.476632
-      - -0.0
-      - -0.0
-    - - 5.2e-05
-      - -0.132985
-      - -0.0
-      - -5.2e-05
+    - - -5.2e-05
       - 0.132985
       - 0.0
-    - - -0.003728
-      - 0.0
-      - -0.136616
-      - 0.003728
+      - 5.2e-05
+      - -0.132985
+      - -0.0
+    - - 0.003728
       - -0.0
       - 0.136616
-    - - 113.756341
+      - -0.003728
+      - 0.0
+      - -0.136616
+    - - -113.756341
+      - -0.0
+      - 0.0
+      - 113.756341
       - 0.0
       - -0.0
-      - -113.756341
-      - -0.0
-      - 0.0
-    - - -0.006777
-      - 0.127989
-      - -0.0
-      - 0.006777
+    - - 0.006777
       - -0.127989
       - 0.0
-    - - -0.005758
+      - -0.006777
+      - 0.127989
       - -0.0
-      - 0.123162
-      - 0.005758
+    - - 0.005758
       - 0.0
       - -0.123162
+      - -0.005758
+      - -0.0
+      - 0.123162
