+++ examples/examples/ru1121_structures/co_diss.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:04:46.252801Z'
+  generated: '2026-03-16T14:10:35.449416Z'
 structure:
   formula: Ru48C2O2
   n_atoms: 52
@@ -67,6 +67,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 4.6.31
+  executed_at: '2012-11-09T08:07:31Z'
   type: frequency
   incar:
     PREC: medium
@@ -114,148 +116,148 @@ vibrations:
     - 52Y
     - 52Z
     matrix:
-    - - -0.300255
-      - 0.088914
-      - -0.254097
-      - 0.002905
-      - 0.000499
-      - 0.000779
-      - -0.483295
-      - -0.139404
-      - -0.007425
-      - -0.001669
-      - -0.001859
-      - 0.000687
-    - - 0.088914
-      - -0.779316
-      - -0.03902
-      - 0.000257
-      - -0.001106
-      - 0.002503
-      - -0.142408
-      - -0.004265
-      - 0.012445
-      - -0.000567
-      - -0.002104
-      - -0.000574
-    - - -0.254097
-      - -0.03902
-      - -0.583199
-      - -0.000629
-      - 0.002592
-      - 0.000646
-      - 0.256879
-      - 0.091933
-      - 0.044599
-      - -0.00068
-      - -0.001128
-      - 0.000296
-    - - 0.002905
-      - 0.000257
-      - -0.000629
-      - -0.30067
-      - 0.089238
-      - -0.255229
-      - -0.002467
-      - -0.001244
-      - -2.6e-05
-      - -0.483135
-      - -0.139654
-      - -0.006954
-    - - 0.000499
-      - -0.001106
-      - 0.002592
-      - 0.089238
-      - -0.781296
-      - -0.038562
-      - -0.000364
-      - -0.002053
-      - -0.000644
-      - -0.14254
-      - -0.004136
-      - 0.012165
-    - - 0.000779
-      - 0.002503
-      - 0.000646
-      - -0.255229
-      - -0.038562
-      - -0.584989
-      - -0.001165
-      - -0.000974
-      - 0.000208
-      - 0.257647
-      - 0.092458
-      - 0.044619
-    - - -0.483295
-      - -0.142408
-      - 0.256879
-      - -0.002467
-      - -0.000364
-      - -0.001165
-      - 0.094479
-      - 0.005632
-      - -0.309712
-      - 0.001225
-      - 0.00022
-      - 0.000181
-    - - -0.139404
-      - -0.004265
-      - 0.091933
-      - -0.001244
-      - -0.002053
-      - -0.000974
-      - 0.005632
-      - -0.414892
-      - -0.210247
-      - 0.000682
-      - -0.001372
-      - 0.000401
-    - - -0.007425
-      - 0.012445
-      - 0.044599
-      - -2.6e-05
-      - -0.000644
-      - 0.000208
-      - -0.309712
-      - -0.210247
-      - -0.757215
-      - 7.6e-05
-      - 0.000553
-      - 0.00178
-    - - -0.001669
-      - -0.000567
-      - -0.00068
-      - -0.483135
-      - -0.14254
-      - 0.257647
-      - 0.001225
-      - 0.000682
-      - 7.6e-05
-      - 0.097854
-      - -0.000743
-      - -0.312021
-    - - -0.001859
-      - -0.002104
-      - -0.001128
-      - -0.139654
-      - -0.004136
-      - 0.092458
-      - 0.00022
-      - -0.001372
-      - 0.000553
-      - -0.000743
-      - -0.409362
-      - -0.209258
-    - - 0.000687
-      - -0.000574
-      - 0.000296
-      - -0.006954
-      - 0.012165
-      - 0.044619
-      - 0.000181
-      - 0.000401
-      - 0.00178
-      - -0.312021
-      - -0.209258
-      - -0.748817
+    - - 0.300255
+      - -0.088914
+      - 0.254097
+      - -0.002905
+      - -0.000499
+      - -0.000779
+      - 0.483295
+      - 0.139404
+      - 0.007425
+      - 0.001669
+      - 0.001859
+      - -0.000687
+    - - -0.088914
+      - 0.779316
+      - 0.03902
+      - -0.000257
+      - 0.001106
+      - -0.002503
+      - 0.142408
+      - 0.004265
+      - -0.012445
+      - 0.000567
+      - 0.002104
+      - 0.000574
+    - - 0.254097
+      - 0.03902
+      - 0.583199
+      - 0.000629
+      - -0.002592
+      - -0.000646
+      - -0.256879
+      - -0.091933
+      - -0.044599
+      - 0.00068
+      - 0.001128
+      - -0.000296
+    - - -0.002905
+      - -0.000257
+      - 0.000629
+      - 0.30067
+      - -0.089238
+      - 0.255229
+      - 0.002467
+      - 0.001244
+      - 2.6e-05
+      - 0.483135
+      - 0.139654
+      - 0.006954
+    - - -0.000499
+      - 0.001106
+      - -0.002592
+      - -0.089238
+      - 0.781296
+      - 0.038562
+      - 0.000364
+      - 0.002053
+      - 0.000644
+      - 0.14254
+      - 0.004136
+      - -0.012165
+    - - -0.000779
+      - -0.002503
+      - -0.000646
+      - 0.255229
+      - 0.038562
+      - 0.584989
+      - 0.001165
+      - 0.000974
+      - -0.000208
+      - -0.257647
+      - -0.092458
+      - -0.044619
+    - - 0.483295
+      - 0.142408
+      - -0.256879
+      - 0.002467
+      - 0.000364
+      - 0.001165
+      - -0.094479
+      - -0.005632
+      - 0.309712
+      - -0.001225
+      - -0.00022
+      - -0.000181
+    - - 0.139404
+      - 0.004265
+      - -0.091933
+      - 0.001244
+      - 0.002053
+      - 0.000974
+      - -0.005632
+      - 0.414892
+      - 0.210247
+      - -0.000682
+      - 0.001372
+      - -0.000401
+    - - 0.007425
+      - -0.012445
+      - -0.044599
+      - 2.6e-05
+      - 0.000644
+      - -0.000208
+      - 0.309712
+      - 0.210247
+      - 0.757215
+      - -7.6e-05
+      - -0.000553
+      - -0.00178
+    - - 0.001669
+      - 0.000567
+      - 0.00068
+      - 0.483135
+      - 0.14254
+      - -0.257647
+      - -0.001225
+      - -0.000682
+      - -7.6e-05
+      - -0.097854
+      - 0.000743
+      - 0.312021
+    - - 0.001859
+      - 0.002104
+      - 0.001128
+      - 0.139654
+      - 0.004136
+      - -0.092458
+      - -0.00022
+      - 0.001372
+      - -0.000553
+      - 0.000743
+      - 0.409362
+      - 0.209258
+    - - -0.000687
+      - 0.000574
+      - -0.000296
+      - 0.006954
+      - -0.012165
+      - -0.044619
+      - -0.000181
+      - -0.000401
+      - -0.00178
+      - 0.312021
+      - 0.209258
+      - 0.748817
   paired_modes_averaged: true
