+++ examples/examples/ru1121_structures/h.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:26.919357Z'
+  generated: '2026-03-16T14:10:28.914948Z'
 structure:
   formula: Ru48H2
   n_atoms: 50
@@ -65,6 +65,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 6.5.1
+  executed_at: '2025-11-02T10:22:59Z'
   type: frequency
   incar:
     LREAL: F
@@ -102,40 +104,40 @@ vibrations:
     - 50Y
     - 50Z
     matrix:
-    - - -0.366554
-      - -0.172162
-      - -0.661046
-      - -0.007244
-      - 0.004996
-      - 0.00269
-    - - -0.203987
-      - -4.646911
-      - 1.406278
-      - 0.000562
-      - 0.011993
-      - 0.003248
-    - - -0.657922
-      - 1.382863
-      - -5.942496
-      - 0.004154
-      - 0.0038
-      - -0.002436
-    - - -0.008157
-      - -0.005323
-      - -0.002525
-      - -0.342887
-      - -0.205963
-      - -0.633805
-    - - 0.000388
-      - 0.014109
-      - 0.001525
-      - -0.207241
-      - -4.605724
-      - 1.41007
-    - - 0.004419
-      - 0.004596
-      - -0.00266
-      - -0.639867
-      - 1.383919
-      - -5.887162
+    - - 0.366554
+      - 0.172162
+      - 0.661046
+      - 0.007244
+      - -0.004996
+      - -0.00269
+    - - 0.203987
+      - 4.646911
+      - -1.406278
+      - -0.000562
+      - -0.011993
+      - -0.003248
+    - - 0.657922
+      - -1.382863
+      - 5.942496
+      - -0.004154
+      - -0.0038
+      - 0.002436
+    - - 0.008157
+      - 0.005323
+      - 0.002525
+      - 0.342887
+      - 0.205963
+      - 0.633805
+    - - -0.000388
+      - -0.014109
+      - -0.001525
+      - 0.207241
+      - 4.605724
+      - -1.41007
+    - - -0.004419
+      - -0.004596
+      - 0.00266
+      - 0.639867
+      - -1.383919
+      - 5.887162
   paired_modes_averaged: true
