+++ examples/examples/ru1121_structures/h2.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:32.767284Z'
+  generated: '2026-03-16T14:10:38.942348Z'
 structure:
   formula: H2
   n_atoms: 2
@@ -17,6 +17,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 6.5.0
+  executed_at: '2025-11-02T08:46:07Z'
   type: frequency
   incar:
     LREAL: F
@@ -57,39 +59,39 @@ vibrations:
     - 2Y
     - 2Z
     matrix:
-    - - -34.097588
+    - - 34.097588
+      - 0.0
+      - 0.0
+      - -34.097588
       - -0.0
       - -0.0
-      - 34.097588
-      - 0.0
-      - 0.0
-    - - 0.013559
-      - 0.007935
-      - 0.0
-      - -0.013559
+    - - -0.013559
       - -0.007935
       - -0.0
-    - - -0.000446
+      - 0.013559
+      - 0.007935
       - 0.0
-      - 0.008395
-      - 0.000446
+    - - 0.000446
       - -0.0
       - -0.008395
-    - - 34.113916
-      - -0.0
-      - -0.0
-      - -34.113916
+      - -0.000446
+      - 0.0
+      - 0.008395
+    - - -34.113916
       - 0.0
       - 0.0
-    - - -0.004671
-      - -0.008489
+      - 34.113916
       - -0.0
-      - 0.004671
+      - -0.0
+    - - 0.004671
       - 0.008489
       - 0.0
-    - - -0.000513
+      - -0.004671
+      - -0.008489
       - -0.0
-      - -0.008399
-      - 0.000513
+    - - 0.000513
       - 0.0
       - 0.008399
+      - -0.000513
+      - -0.0
+      - -0.008399
