+++ examples/examples/ru1121_structures/h2o__2.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:33.436734Z'
+  generated: '2026-03-16T14:10:39.426460Z'
 structure:
   formula: OH2
   n_atoms: 3
@@ -18,6 +18,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 6.5.0
+  executed_at: '2025-11-02T08:44:01Z'
   type: frequency
   incar:
     LREAL: F
@@ -65,84 +67,84 @@ vibrations:
     - 3Y
     - 3Z
     matrix:
-    - - 0.147426
-      - 0.0
-      - 0.004186
-      - -0.073713
-      - 0.000906
-      - -0.002093
-      - -0.073713
+    - - -0.147426
+      - -0.0
+      - -0.004186
+      - 0.073713
       - -0.000906
-      - -0.002093
-    - - 0.0
-      - -61.674544
-      - 0.001474
-      - -0.0
-      - 30.837401
-      - -24.256773
-      - -0.0
-      - 30.837143
-      - 24.255299
+      - 0.002093
+      - 0.073713
+      - 0.000906
+      - 0.002093
     - - -0.0
+      - 61.674544
+      - -0.001474
+      - 0.0
+      - -30.837401
+      - 24.256773
+      - 0.0
+      - -30.837143
+      - -24.255299
+    - - 0.0
+      - 0.0
+      - 41.042586
       - -0.0
-      - -41.042586
-      - 0.0
-      - -18.409295
-      - 20.521293
-      - 0.0
       - 18.409295
-      - 20.521293
-    - - 0.068262
-      - -0.000274
-      - -0.000396
-      - -0.027548
-      - -0.000185
-      - 0.0001
-      - -0.040714
-      - 0.000459
-      - 0.000296
-    - - 0.0
-      - 30.961351
-      - -18.481229
+      - -20.521293
       - -0.0
-      - -33.929567
-      - 21.370888
-      - -0.0
-      - 2.968215
-      - -2.889658
+      - -18.409295
+      - -20.521293
+    - - -0.068262
+      - 0.000274
+      - 0.000396
+      - 0.027548
+      - 0.000185
+      - -0.0001
+      - 0.040714
+      - -0.000459
+      - -0.000296
     - - -0.0
-      - -24.213595
-      - 20.551567
+      - -30.961351
+      - 18.481229
       - 0.0
-      - 21.320612
-      - -19.708165
+      - 33.929567
+      - -21.370888
       - 0.0
-      - 2.892983
-      - -0.843402
-    - - 0.06309
-      - -0.00045
-      - 0.004578
-      - -0.037554
-      - 0.003057
-      - -0.002113
-      - -0.025535
-      - -0.002606
-      - -0.002465
+      - -2.968215
+      - 2.889658
     - - 0.0
-      - 30.960124
-      - 18.47947
+      - 24.213595
+      - -20.551567
       - -0.0
-      - 2.965817
-      - 2.889956
+      - -21.320612
+      - 19.708165
       - -0.0
-      - -33.925941
-      - -21.369426
-    - - 0.0
-      - 24.201835
-      - 20.544209
-      - -0.0
-      - -2.889827
-      - -0.839736
+      - -2.892983
+      - 0.843402
+    - - -0.06309
+      - 0.00045
+      - -0.004578
+      - 0.037554
+      - -0.003057
+      - 0.002113
+      - 0.025535
+      - 0.002606
+      - 0.002465
+    - - -0.0
+      - -30.960124
+      - -18.47947
       - 0.0
-      - -21.312008
-      - -19.704473
+      - -2.965817
+      - -2.889956
+      - 0.0
+      - 33.925941
+      - 21.369426
+    - - -0.0
+      - -24.201835
+      - -20.544209
+      - 0.0
+      - 2.889827
+      - 0.839736
+      - -0.0
+      - 21.312008
+      - 19.704473
