+++ examples/examples/ru1121_structures/o.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:28.998466Z'
+  generated: '2026-03-16T14:10:30.374782Z'
 structure:
   formula: Ru48O2
   n_atoms: 50
@@ -65,6 +65,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 4.6.31
+  executed_at: '2013-01-09T08:22:13Z'
   type: frequency
   incar:
     PREC: medium
@@ -101,40 +103,40 @@ vibrations:
     - 50Y
     - 50Z
     matrix:
-    - - -0.084287
-      - 0.072992
-      - -0.008534
-      - -0.000661
-      - -0.001122
-      - -8.6e-05
-    - - 0.072992
-      - -0.616145
-      - -0.285505
-      - -6.4e-05
-      - -0.000316
-      - -0.000956
-    - - -0.008534
-      - -0.285505
-      - -1.159589
-      - 0.003462
-      - -0.001188
-      - 0.001586
-    - - -0.000661
-      - -6.4e-05
-      - 0.003462
-      - -0.082672
-      - 0.067378
-      - -0.011762
-    - - -0.001122
-      - -0.000316
-      - -0.001188
-      - 0.067378
-      - -0.612596
-      - -0.276554
-    - - -8.6e-05
-      - -0.000956
-      - 0.001586
-      - -0.011762
-      - -0.276554
-      - -1.159098
+    - - 0.084287
+      - -0.072992
+      - 0.008534
+      - 0.000661
+      - 0.001122
+      - 8.6e-05
+    - - -0.072992
+      - 0.616145
+      - 0.285505
+      - 6.4e-05
+      - 0.000316
+      - 0.000956
+    - - 0.008534
+      - 0.285505
+      - 1.159589
+      - -0.003462
+      - 0.001188
+      - -0.001586
+    - - 0.000661
+      - 6.4e-05
+      - -0.003462
+      - 0.082672
+      - -0.067378
+      - 0.011762
+    - - 0.001122
+      - 0.000316
+      - 0.001188
+      - -0.067378
+      - 0.612596
+      - 0.276554
+    - - 8.6e-05
+      - 0.000956
+      - -0.001586
+      - 0.011762
+      - 0.276554
+      - 1.159098
   paired_modes_averaged: true
