+++ examples/examples/ru1121_structures/o_hydr.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:04:49.663481Z'
+  generated: '2026-03-16T14:10:37.645805Z'
 structure:
   formula: Ru48O2H2
   n_atoms: 52
@@ -67,6 +67,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 4.6.31
+  executed_at: '2012-08-14T19:09:34Z'
   type: frequency
   incar:
     PREC: medium
@@ -114,148 +116,148 @@ vibrations:
     - 52Y
     - 52Z
     matrix:
-    - - 0.052907
-      - 0.060238
-      - 0.103899
-      - -0.000107
-      - -0.000266
-      - 0.000471
-      - -0.588302
-      - -0.177032
-      - -0.754648
-      - 0.001332
-      - 0.000549
-      - 0.001473
-    - - 0.060238
-      - -0.335469
-      - -0.262594
-      - -0.000522
-      - -0.000651
-      - -0.0003
-      - 0.045487
-      - -0.099
-      - -0.001203
-      - 0.000373
-      - 0.001357
-      - 0.000523
-    - - 0.103899
-      - -0.262594
-      - -0.939577
-      - -0.000232
-      - -0.000492
-      - -5.3e-05
-      - 0.539423
-      - 0.103422
-      - 0.176982
-      - 0.003093
-      - -0.000706
-      - 0.000226
-    - - -0.000107
-      - -0.000522
-      - -0.000232
-      - 0.052744
-      - 0.064051
-      - 0.110953
-      - -0.001392
-      - 0.000407
-      - 0.000818
-      - -0.588957
-      - -0.17667
-      - -0.753076
-    - - -0.000266
-      - -0.000651
-      - -0.000492
-      - 0.064051
-      - -0.338944
-      - -0.267634
-      - -0.002309
-      - 0.00171
-      - 0.000995
-      - 0.046137
-      - -0.098451
-      - -0.000371
-    - - 0.000471
-      - -0.0003
-      - -5.3e-05
-      - 0.110953
-      - -0.267634
-      - -0.940818
-      - -0.002639
-      - -0.00011
-      - -0.000403
-      - 0.538376
-      - 0.103497
-      - 0.178537
-    - - -0.588302
-      - 0.045487
-      - 0.539423
-      - -0.001392
-      - -0.002309
-      - -0.002639
-      - 1.339496
-      - 0.182766
-      - 1.407709
-      - -0.011559
-      - 0.005759
-      - 0.001654
-    - - -0.177032
-      - -0.099
-      - 0.103422
-      - 0.000407
-      - 0.00171
-      - -0.00011
-      - 0.182766
-      - -0.589412
-      - -0.35084
-      - 0.003451
-      - -0.011194
-      - 0.000197
-    - - -0.754648
-      - -0.001203
-      - 0.176982
-      - 0.000818
-      - 0.000995
-      - -0.000403
-      - 1.407709
-      - -0.35084
-      - -4.627834
-      - -0.008622
-      - 0.001206
-      - -0.013641
-    - - 0.001332
-      - 0.000373
-      - 0.003093
-      - -0.588957
-      - 0.046137
-      - 0.538376
-      - -0.011559
-      - 0.003451
-      - -0.008622
-      - 1.341039
-      - 0.165903
-      - 1.402726
-    - - 0.000549
-      - 0.001357
-      - -0.000706
-      - -0.17667
-      - -0.098451
-      - 0.103497
-      - 0.005759
-      - -0.011194
-      - 0.001206
-      - 0.165903
-      - -0.587069
-      - -0.346732
-    - - 0.001473
-      - 0.000523
-      - 0.000226
-      - -0.753076
-      - -0.000371
-      - 0.178537
-      - 0.001654
-      - 0.000197
-      - -0.013641
-      - 1.402726
-      - -0.346732
-      - -4.628795
+    - - -0.052907
+      - -0.060238
+      - -0.103899
+      - 0.000107
+      - 0.000266
+      - -0.000471
+      - 0.588302
+      - 0.177032
+      - 0.754648
+      - -0.001332
+      - -0.000549
+      - -0.001473
+    - - -0.060238
+      - 0.335469
+      - 0.262594
+      - 0.000522
+      - 0.000651
+      - 0.0003
+      - -0.045487
+      - 0.099
+      - 0.001203
+      - -0.000373
+      - -0.001357
+      - -0.000523
+    - - -0.103899
+      - 0.262594
+      - 0.939577
+      - 0.000232
+      - 0.000492
+      - 5.3e-05
+      - -0.539423
+      - -0.103422
+      - -0.176982
+      - -0.003093
+      - 0.000706
+      - -0.000226
+    - - 0.000107
+      - 0.000522
+      - 0.000232
+      - -0.052744
+      - -0.064051
+      - -0.110953
+      - 0.001392
+      - -0.000407
+      - -0.000818
+      - 0.588957
+      - 0.17667
+      - 0.753076
+    - - 0.000266
+      - 0.000651
+      - 0.000492
+      - -0.064051
+      - 0.338944
+      - 0.267634
+      - 0.002309
+      - -0.00171
+      - -0.000995
+      - -0.046137
+      - 0.098451
+      - 0.000371
+    - - -0.000471
+      - 0.0003
+      - 5.3e-05
+      - -0.110953
+      - 0.267634
+      - 0.940818
+      - 0.002639
+      - 0.00011
+      - 0.000403
+      - -0.538376
+      - -0.103497
+      - -0.178537
+    - - 0.588302
+      - -0.045487
+      - -0.539423
+      - 0.001392
+      - 0.002309
+      - 0.002639
+      - -1.339496
+      - -0.182766
+      - -1.407709
+      - 0.011559
+      - -0.005759
+      - -0.001654
+    - - 0.177032
+      - 0.099
+      - -0.103422
+      - -0.000407
+      - -0.00171
+      - 0.00011
+      - -0.182766
+      - 0.589412
+      - 0.35084
+      - -0.003451
+      - 0.011194
+      - -0.000197
+    - - 0.754648
+      - 0.001203
+      - -0.176982
+      - -0.000818
+      - -0.000995
+      - 0.000403
+      - -1.407709
+      - 0.35084
+      - 4.627834
+      - 0.008622
+      - -0.001206
+      - 0.013641
+    - - -0.001332
+      - -0.000373
+      - -0.003093
+      - 0.588957
+      - -0.046137
+      - -0.538376
+      - 0.011559
+      - -0.003451
+      - 0.008622
+      - -1.341039
+      - -0.165903
+      - -1.402726
+    - - -0.000549
+      - -0.001357
+      - 0.000706
+      - 0.17667
+      - 0.098451
+      - -0.103497
+      - -0.005759
+      - 0.011194
+      - -0.001206
+      - -0.165903
+      - 0.587069
+      - 0.346732
+    - - -0.001473
+      - -0.000523
+      - -0.000226
+      - 0.753076
+      - 0.000371
+      - -0.178537
+      - -0.001654
+      - -0.000197
+      - 0.013641
+      - -1.402726
+      - 0.346732
+      - 4.628795
   paired_modes_averaged: true
