+++ examples/examples/ru1121_structures/oh.yaml
@@ -1,6 +1,6 @@
 pymkmkit:
   version: 0.1.0
-  generated: '2026-02-23T19:05:30.071160Z'
+  generated: '2026-03-16T14:10:31.122392Z'
 structure:
   formula: Ru48O2H2
   n_atoms: 52
@@ -67,6 +67,8 @@ structure:
   - true
 calculation:
   code: VASP
+  version: 6.5.1
+  executed_at: '2025-11-02T10:26:06Z'
   type: frequency
   incar:
     LREAL: F
@@ -114,148 +116,148 @@ vibrations:
     - 52Y
     - 52Z
     matrix:
-    - - -38.677007
-      - -0.932893
-      - 18.256741
-      - -0.000267
-      - -0.021214
-      - -0.003312
-      - 37.305766
-      - 1.648667
-      - -17.370651
-      - 0.004833
-      - 0.000894
-      - -0.000649
-    - - -0.942642
-      - -6.041602
-      - -1.955329
-      - 0.010413
-      - -0.00722
-      - -0.003788
-      - 1.508117
-      - 1.422419
-      - -0.731755
-      - -0.002054
-      - -0.001007
-      - -0.003383
-    - - 18.182172
-      - -2.03556
-      - -18.40952
-      - 0.008465
-      - 0.014495
-      - 0.056601
-      - -18.111126
-      - -0.908326
-      - 9.21249
-      - -0.007101
-      - -0.001071
-      - -0.023353
-    - - 0.025742
-      - -0.01221
-      - 0.027344
-      - -38.758154
-      - -0.946811
-      - 18.236066
-      - -0.004681
-      - 0.002906
-      - 0.005591
-      - 37.30483
-      - 1.651304
-      - -17.368691
-    - - 0.008298
-      - -0.002706
-      - 0.002021
-      - -0.97506
-      - -6.105231
-      - -2.020297
-      - -0.000269
-      - 0.000339
-      - -0.003446
-      - 1.513543
-      - 1.423092
-      - -0.73155
-    - - 0.009166
-      - 0.005001
-      - 0.052604
-      - 18.162828
-      - -2.082045
-      - -18.38423
-      - -0.006028
-      - -0.000964
-      - -0.022012
-      - -18.105987
-      - -0.905303
-      - 9.209394
-    - - 37.22498
-      - 1.470469
-      - -18.180763
-      - 0.023457
-      - 0.003841
-      - 0.010211
-      - -37.45748
-      - -1.352769
-      - 18.255873
-      - -0.010587
-      - 0.000599
-      - 0.011519
-    - - 1.684235
-      - 1.392697
-      - -0.95398
-      - -0.009555
-      - 0.018358
-      - 0.007566
-      - -1.376782
-      - -1.644566
-      - 0.637705
-      - 0.004666
-      - 0.00297
-      - -0.002439
-    - - -17.335799
-      - -0.717162
-      - 9.207415
-      - -0.003157
-      - 0.004216
-      - -0.006733
-      - 18.214861
-      - 0.614015
-      - -10.219577
-      - 0.000841
-      - 0.001233
-      - 0.005942
-    - - 0.012615
-      - 0.010703
-      - 0.014951
-      - 37.291511
-      - 1.497915
-      - -18.204882
-      - 0.011593
-      - -0.001726
-      - -0.018551
-      - -37.540818
-      - -1.355171
-      - 18.270833
-    - - -0.003459
-      - 0.008907
-      - -0.003973
-      - 1.658908
-      - 1.388468
-      - -0.941388
-      - 0.003981
-      - 0.002824
-      - -0.000606
-      - -1.366849
-      - -1.677795
-      - 0.643756
-    - - 0.004835
-      - -0.010179
-      - -0.026664
-      - -17.329225
-      - -0.737495
-      - 9.170669
-      - -0.002268
-      - 0.000156
-      - 0.008462
-      - 18.214075
-      - 0.621717
-      - -10.239149
+    - - 38.677007
+      - 0.932893
+      - -18.256741
+      - 0.000267
+      - 0.021214
+      - 0.003312
+      - -37.305766
+      - -1.648667
+      - 17.370651
+      - -0.004833
+      - -0.000894
+      - 0.000649
+    - - 0.942642
+      - 6.041602
+      - 1.955329
+      - -0.010413
+      - 0.00722
+      - 0.003788
+      - -1.508117
+      - -1.422419
+      - 0.731755
+      - 0.002054
+      - 0.001007
+      - 0.003383
+    - - -18.182172
+      - 2.03556
+      - 18.40952
+      - -0.008465
+      - -0.014495
+      - -0.056601
+      - 18.111126
+      - 0.908326
+      - -9.21249
+      - 0.007101
+      - 0.001071
+      - 0.023353
+    - - -0.025742
+      - 0.01221
+      - -0.027344
+      - 38.758154
+      - 0.946811
+      - -18.236066
+      - 0.004681
+      - -0.002906
+      - -0.005591
+      - -37.30483
+      - -1.651304
+      - 17.368691
+    - - -0.008298
+      - 0.002706
+      - -0.002021
+      - 0.97506
+      - 6.105231
+      - 2.020297
+      - 0.000269
+      - -0.000339
+      - 0.003446
+      - -1.513543
+      - -1.423092
+      - 0.73155
+    - - -0.009166
+      - -0.005001
+      - -0.052604
+      - -18.162828
+      - 2.082045
+      - 18.38423
+      - 0.006028
+      - 0.000964
+      - 0.022012
+      - 18.105987
+      - 0.905303
+      - -9.209394
+    - - -37.22498
+      - -1.470469
+      - 18.180763
+      - -0.023457
+      - -0.003841
+      - -0.010211
+      - 37.45748
+      - 1.352769
+      - -18.255873
+      - 0.010587
+      - -0.000599
+      - -0.011519
+    - - -1.684235
+      - -1.392697
+      - 0.95398
+      - 0.009555
+      - -0.018358
+      - -0.007566
+      - 1.376782
+      - 1.644566
+      - -0.637705
+      - -0.004666
+      - -0.00297
+      - 0.002439
+    - - 17.335799
+      - 0.717162
+      - -9.207415
+      - 0.003157
+      - -0.004216
+      - 0.006733
+      - -18.214861
+      - -0.614015
+      - 10.219577
+      - -0.000841
+      - -0.001233
+      - -0.005942
+    - - -0.012615
+      - -0.010703
+      - -0.014951
+      - -37.291511
+      - -1.497915
+      - 18.204882
+      - -0.011593
+      - 0.001726
+      - 0.018551
+      - 37.540818
+      - 1.355171
+      - -18.270833
+    - - 0.003459
+      - -0.008907
+      - 0.003973
+      - -1.658908
+      - -1.388468
+      - 0.941388
+      - -0.003981
+      - -0.002824
+      - 0.000606
+      - 1.366849
+      - 1.677795
+      - -0.643756
+    - - -0.004835
+      - 0.010179
+      - 0.026664
+      - 17.329225
+      - 0.737495
+      - -9.170669
+      - 0.002268
+      - -0.000156
+      - -0.008462
+      - -18.214075
+      - -0.621717
+      - 10.239149
   paired_modes_averaged: true
